% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateMap.R
\name{stateMap}
\alias{stateMap}
\title{State level thematic map}
\usage{
stateMap(
  data = NULL,
  parameter = NULL,
  state_SPDF = "USCensusStates_02",
  palette = "YlOrBr",
  breaks = NULL,
  conusOnly = TRUE,
  stateCode = NULL,
  projection = NULL,
  stateBorderColor = "gray50",
  title = NULL
)
}
\arguments{
\item{data}{Dataframe containing values to plot a state map. This dataframe
must contain a column named \code{stateCode} with the 2-character state code.}

\item{parameter}{Name of the column of data in \code{state_SPDF} to use fo
coloring the map.}

\item{state_SPDF}{SpatialPolygonsDataFrame with US States.}

\item{palette}{A palette name or a vector of colors based on RColorBrewer.}

\item{breaks}{Numeric vector of break points. Must be 1 greater than \code{n}.}

\item{conusOnly}{Logical specifying CONtinental US state codes.}

\item{stateCode}{Vector of state codes.}

\item{projection}{Specified method to represent surface of Earth.}

\item{stateBorderColor}{The color of the state border to display.}

\item{title}{Text string to use as the plot title.}
}
\value{
A ggplot object.
}
\description{
Uses the \pkg{tmap} package to generate a thematic map at the
county level. Input consists of a dataframe with county identifiers.
}
\details{
See \code{tmap::tm_fill()} for a more detailed description of
the following parameters:

\itemize{
\item{\code{palette}}
\item{\code{n}}
\item{\code{breaks}}
}
}
\examples{
\donttest{
library(MazamaSpatialPlots)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("USCensusStates_02")

stateMap(
  data = example_US_stateObesity, 
  parameter = "obesityRate", 
  state_SPDF = "USCensusStates_02",
  palette = "BuPu",
  stateBorderColor = "white",
  title = "2018 Obesity by State"
)
}
}
