% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countyMap.R
\name{countyMap}
\alias{countyMap}
\title{County level thematic map}
\usage{
countyMap(
  data = NULL,
  countySPDF = "USCensusCounties",
  stateSPDF = "USCensusStates",
  paletteName = "YlOrBr",
  breaks = NULL,
  conusOnly = TRUE,
  stateCode = NULL,
  projection = NULL,
  stateBorderColor = "gray50",
  countyBorderColor = "white"
)
}
\arguments{
\item{data}{TODO}

\item{countySPDF}{TODO}

\item{stateSPDF}{TODO}

\item{paletteName}{TODO}

\item{breaks}{TODO}

\item{conusOnly}{TODO}

\item{stateCode}{TODO}

\item{projection}{TODO}

\item{stateBorderColor}{TODO}

\item{countyBorderColor}{TODO}
}
\value{
A ggplot object.
}
\description{
Uses the \pkg{tmap} package to generate a thematic map at the
county level. Input consists of a dataframe with county identifiers.
}
\examples{
\donttest{
library(MazamaSpatialPlots)

# TODO:  Load package internal example_countyDF

countyMap(countyDF, "USCensusCounties_05")

}
}
