% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countyMap.R
\name{countyMap}
\alias{countyMap}
\title{County level thematic map}
\usage{
countyMap(
  data = NULL,
  parameter = NULL,
  state_SPDF = "USCensusStates_02",
  county_SPDF = "USCensusCounties_02",
  palette = "YlOrBr",
  breaks = NULL,
  conusOnly = TRUE,
  stateCode = NULL,
  projection = NULL,
  stateBorderColor = "gray50",
  countyBorderColor = "white",
  title = NULL
)
}
\arguments{
\item{data}{Dataframe containing values to plot. This dataframe
must contain a column named \code{countyFIPS} with the 5-digit FIPS code.}

\item{parameter}{Name of the column of data in \code{data} to use for
coloring the map.}

\item{state_SPDF}{SpatialPolygonsDataFrame with US states. It's data 
\code{@slot} must contain a column named \code{stateCode}}

\item{county_SPDF}{SpatialPolygonsDataFrame with US counties. It's data 
\code{@slot} must contain a column named \code{stateCode} and \code{countyFIPS}}

\item{palette}{Palette name or a vector of colors based on RColorBrewer.}

\item{breaks}{Numeric vector of break points.}

\item{conusOnly}{Logical specifying Continental US state codes. Ignored when 
the \code{stateCode} argument is specified.}

\item{stateCode}{Vector of state codes to include on the map.}

\item{projection}{Specified method to represent surface of Earth.}

\item{stateBorderColor}{Color used for state borders.}

\item{countyBorderColor}{Color used for county borders.}

\item{title}{Text string to use as the plot title.}
}
\value{
A ggplot object.
}
\description{
Uses the \pkg{tmap} package to generate a thematic map at the
county level. Input consists of a dataframe with \code{countyFIPS} identifiers.
}
\details{
See \code{tmap::tm_fill()} for a more detailed description of
the following parameters:

\itemize{
\item{\code{palette}}
\item{\code{breaks}}
}
}
\examples{
\donttest{
library(MazamaSpatialPlots)
mazama_initialize()

countyMap(
  data = example_US_countyCovid, 
  parameter = "cases",
  breaks = c(0,100,200,500,1000,2000,5000,10000,20000,50000,1e6),
  title = "Covid cases by county -- June 01, 2020"
)
}
}
