% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateMap.R
\name{stateMap}
\alias{stateMap}
\title{State level thematic map}
\usage{
stateMap(
  data = NULL,
  state_SPDF = "USCensusStates",
  paletteName = "YlOrBr",
  breaks = NULL,
  conusOnly = TRUE,
  stateCode = NULL,
  projection = NULL,
  stateBorderColor = "gray50"
)
}
\arguments{
\item{data}{A dataframe containing values to plot a county map. At a minimum
there will be stateCode, countyName, and countyFIPS.}

\item{state_SPDF}{ector of US States.}

\item{paletteName}{A palette name or a vector of colors based on RColorBrewer.}

\item{breaks}{If style is fixed, breaks can be specified.}

\item{conusOnly}{Logical specifying CONtinental US state codes.}

\item{stateCode}{Vector of state codes.}

\item{projection}{Specified method to represent surface of Earth.}

\item{stateBorderColor}{The color of the state border to display.}
}
\value{
A ggplot object.
}
\description{
Uses the \pkg{tmap} package to generate a thematic map at the
county level. Input consists of a dataframe with county identifiers.
}
\examples{
\donttest{
library(MazamaSpatialPlots)

stateMap(stateDF, "USCensusCounties_05")

}
}
